fix implementation for symmetric problem
It is recommended to use build_problem("PROBLEM_NAME",true) instead of build problem. Second argument is set to true if square structurally symmetric matrix , i.e. most of the cases will be in that category.

Matrix can be nonsysmetric, but on rows and columns have to have the same fields. Elements can be non-symmetric in that case.